+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to
 - Deprecation notes in the `Error` module for `Resolution`, `CollectErrsT` and
   `runCollectErrs`. These symbols are part of the old executor and aren't used
   anymore, it will be deprecated in the future and removed.
+- `TH` module with the `gql` quasi quoter.
 
 ### Fixed
 - Error messages are more concrete, they also contain type information and